Day 3 exercise notebook
We tried to import the csv data to QGIS. QGIS interpreted the lat longs as metres away from the north pole. So, we used ogr2ogr to make it into a geojson file. This successfully made the points appear on the map, however, they were in Russia rather than in Greenland. But then it was lunchtime, so here's our notebook.