Day Three symbology exercise
These files are the symbology styles from our data set (the DEM) and the data set that we chose to overlay on ours (the thick billed murre data set from QGreenland). The dataset that we had been using in the previous exercise from Group Four, the stream data, we were still unable to have projected in Greenland and it was still showing up in Russia, so we used a QGreenland dataset for this exercise.